Data Ingestion/requirements.txt
aiokafka==0.10.0
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-cache2==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
kafka-python==1.3.5
packaging==23.2
pendulum==3.0.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
websockets==12.0
